requirements.txt
playwright==1.40.0
beautifulsoup4==4.12.2
pandas==2.2.3
sqlalchemy==2.0.22
psycopg2-binary==2.9.7
luigi==3.0.1
python-dotenv==1.0.0